+++ kkanbu/operation/tests.py
@@ -1,0 +1,6 @@
+from django.test import TestCase
+
+
+class OperationTest(TestCase):
+    def test_post_blame(self):
+        pass